chore: update test cases
